Renaming find{Nodes|Links}2.py to find{Nodes|Links}.py
