+++ tangelo/mongo/web/findNodes.py
@@ -4,20 +4,20 @@ import json
 from pymongo import MongoClient
 
 
-def run(host=None, db=None, coll=None, spec=None, singleton=json.dumps(False)):
+def run(host=None, db=None, coll=None, spec=None):
     # Connect to the mongo collection.
     client = MongoClient(host)
     db = client[db]
     graph = db[coll]
 
-    spec = bson.json_util.loads(spec)
-    singleton = json.loads(singleton)
+    spec = json.loads(spec)
+    spec2 = {"type": "node"}
+    for k, v in spec.iteritems():
+        if k == "key":
+            spec2["_id"] = ObjectId(v)
+        else:
+            spec2["data.%s" % (k)] = v
 
-    spec.update({"type": "node"})
-
-    if singleton:
-        result = graph.find_one(spec)
-    else:
-        result = list(graph.find(spec))
+    result = list(graph.find(spec2))
 
     return bson.json_util.dumps(result)
